preliminary support for windowing
git-svn-id: https://svn.osgeo.org/gdal@16942 f0d54148-0727-0410-94bb-9a71ac55c965

+++ sandbox/warmerdam/pcidsk/src/cpcidskgeoref.cpp
@@ -66,7 +66,7 @@ void CPCIDSKGeoref::Load()
 /* -------------------------------------------------------------------- */
 /*      Load the segment contents into a buffer.                        */
 /* -------------------------------------------------------------------- */
-    seg_data.SetSize( data_size - 1024 );
+    seg_data.SetSize( (int) (data_size - 1024) );
 
     ReadFromFile( seg_data.buffer, 0, data_size - 1024 );
